@@ -13,7 +13,7 @@
 #  updated_at :datetime         not null
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :help_text do
     category { Faker::Lorem.word }
     help_type { Faker::Lorem.word }
