@@ -26,8 +26,8 @@ RSpec.configure do |config|
   end
 
   # add setting
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
   config.before(:all) do
-    FactoryGirl.reload
+    FactoryBot.reload
   end
 end
